i believe the librabry has been appropriately converted to using skiasharp
customcanvasview is now the standard interface into the low poly library.
-moved customcanvasview to the shared project since it now holds refrences to low poly lib objects now
-animations are posted to the view and will be handled accordingly

currently the animations are not rendering correctly but they are close.
-frame currently being updated  and writing to console
- touch animation draws red circle

bitmap pool longer needed since a reference to 1 canvas is now being passed around

animation rendering cycle changing. instead of generating bitmaps and timing their display on screen, a reference to the same canvas is passed around and drawn on