fixed library bug in build
added dataflowex and its dependencies to the ios project in order for the app to be able to run. Looks like lowpolylib isnt exporting its depencies so they have to be readded to the android and ios projects. 

changed version of andriod views lib to match poly lib project

made animation update view transparent